+++ dependencies/campello_gpu.cmake
@@ -6,7 +6,7 @@ include(FetchContent)
 FetchContent_Declare(
     campello_gpu
     GIT_REPOSITORY https://github.com/rusoleal/campello_gpu
-    GIT_TAG        v0.7.0
+    GIT_TAG        v0.8.0
 )
 
 if(NOT campello_gpu_POPULATED)