+++ dependencies/vector_math.cmake
@@ -6,7 +6,7 @@ include(FetchContent)
 FetchContent_Declare(
     vector_math
     GIT_REPOSITORY https://github.com/rusoleal/vector_math
-    GIT_TAG        main
+    GIT_TAG        v0.3.5
 )
 
 if(NOT vector_math_POPULATED)